Strengthens regex for jicofo secrets
The jitsi-meet installation automatically generates secrets for the
jicofo and videobridge components. Those secrets are then reused by the
role to ensure idempotence. The logic used to generate the secret
strings is opaque to me for the moment (would have to check the
Java source code), but on a recent test run, the serverspec check for
jicofo choked on an '@' for the secret value. Updated regex accordingly,
and left a comment in the videobridge serverspec test file, which uses
a similar regex and therefore may need to be updated in the future also.