Uses prosodyctl for cert and auth management
Rather than reverse engineer the OpenSSL generation and Jicofo
registration subcommands for `prosodyctl`, use the tool provided and
skip the task if resultant files are created already.

There's a slight drawback in that secrets can't be trivially rotated: if
you update the secrets used between Jitsi components, the prosodyctl
task will be skipped on a subsequent playbook run because it checks for
file existence, not file content. That's OK, because the vars are
dynamically inferred anyway, and not exposed in role defaults. It'd be
trivial to implement a force override boolean that regenerates the cert.

Regarding the 'replace' filter, the looks verbose, but there's a reason.
The unintuitive URL escaping of directory names performed by the
prosodyctl command makes testing for changed state a bit tricky.
The '%2e' string must appear literally in the first portion of the
'auth.<fqdn>' string, since the |replace filter operates solely on the
fqdn portion of the var, not on the concatenated result.