Consolidates apt repo configuration tasks
Deciding between stable/unstable just got simpler, since the Jitsi team
recently started signing both the unstable and stable releases with the
same key. That means we can use the same setup task to configure the
key, and only need to customize the repo name anymore.

Still using a dict-based vars structure to extract what we want, but it
works.