Makes Nginx configuration optional
The role provides the sane default of a preconfigured Nginx vhost,
created following the Jitsi Meet project docs. That's fine for a
baseline configuration, but makes managing for the Jitsi service via
a separate role quite difficult. Let's make the Nginx configuration
opt-out, so it still runs by default but can be skipped if admins
prefer another method for managing vhosts.

The vhost template still ships with the role and should be consulted
when writing a custom vhost config.

Closes #28.